Refactored MovesStrategies into their own directory. Initialized path strategies.
