+++ include/componentsystems/strategies/moves/SteadyMovesStrategy.h
@@ -1,7 +1,7 @@
 #ifndef STEADYMOVESSTRATEGY_H
 #define STEADYMOVESSTRATEGY_H
 
-#include "componentsystems/strategies/MovesStrategy.h"
+#include "componentsystems/strategies/moves/MovesStrategy.h"
 
 namespace TIE {
