+++ src/componentsystems/MovesComponentSystem.cpp
@@ -8,9 +8,9 @@
 #include "componentsystems/GridComponentSystem.h"
 #include "componentsystems/MessagesComponentSystem.h"
 #include "componentsystems/WireframeComponentSystem.h"
-#include "componentsystems/strategies/AcceleratedTurnMovesStrategy.h"
-#include "componentsystems/strategies/SnapTurnMovesStrategy.h"
-#include "componentsystems/strategies/SteadyMovesStrategy.h"
+#include "componentsystems/strategies/moves/AcceleratedTurnMovesStrategy.h"
+#include "componentsystems/strategies/moves/SnapTurnMovesStrategy.h"
+#include "componentsystems/strategies/moves/SteadyMovesStrategy.h"
 #include "managers/ComponentSystemsManager.h"
 #include "managers/MessageManager.h"
 #include "managers/ScriptManager.h"
