+++ src/componentsystems/strategies/path/PathStrategy.cpp
@@ -1,0 +1,7 @@
+#include "componentsystems/strategies/path/PathStrategy.h"
+
+using namespace TIE;
+
+bool PathStrategy::execute(const float delta, TIEntity& tientity) { 
+    return false;
+}
